refactor(useValidation): remove unnecessary watch clear and use of lodash's get
* lodash get is not necessary because Javascript's ES6-era syntax already allows nullish coalescing
* There's no need to handle the clearage of the watcher sin Vue garbage collects it inside an effect scope
