+++ src/pages/login/index.vue
@@ -50,7 +50,7 @@ import { z } from 'zod';
 import { UsersApi } from '@/api';
 import { useApi } from '@/composables/apis';
 import { auth } from '@/store/auth';
-import useValidation from '@/utils/useValidation';
+import { useValidation } from '@/composables/use-validation';
 
 const router = useRouter();
